combineVersion/piCode/requirement.txt
adafruit-blinka
RPi.GPIO
borad
adafruit-circuitpython-sgp30
paho-mqtt
vpype
serial